+++ CHANGELOG.md
@@ -6,3 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.1] - 2019-06-27
+### Changed
+- ApiSave renamed to ApiSaveData
+
+## [1.0.0] - 2019-06-26
+### Added
+- Project inited
+- API Handler
+- Tests